# crate/server/documentation/openapi.yaml
...
openapi: 3.1.0
info:
  title: Cosmian KMS
  version: 5.0.0
  description: |
    High-performance, FIPS 140-3 compliant Key Management System implementing **KMIP 2.1 and 1.4** over HTTP/TLS.

    ## Authentication

    Most endpoints require one of the following authentication methods:

    - **Bearer JWT** — pass a JWT from your OIDC provider in `Authorization: Bearer <token>`
    - **API Token** — pass an API token in `Authorization: Bearer <token>` (server-issued opaque token)
    - **Mutual TLS (mTLS)** — present a client certificate during the TLS handshake

    Public endpoints (`/health`, `/version`, `/swagger`, `/openapi.yaml`) require no authentication.
servers:
  - url: /
    description: Current server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by your OIDC provider, or an opaque API token issued by the KMS server
    mtls:
      type: mutualTLS
      description: Client certificate presented during TLS handshake
    vaultToken:
      type: apiKey
      in: header
      name: X-Vault-Token
      description: |
        Vault-compatible bearer token issued by the Cosmian auth-verifier.
        Required on all `/v1/transit/*` and `/v1/pki/*` endpoints when
        `vault_api_enabled = true`.
    # ── REST Crypto tag management ───────────────────────────────────────────

    TagsRequest:
      description: Request body for adding or removing user tags on a key.
      type: object
      required: [tags]
      properties:
        tags:
          type: array
          items:
            type: string
          description: >
            Tags to add or remove. Each tag must be a non-empty string and must
            not start with `_` (that prefix is reserved for system tags).
          example: [jwks, prod]

    TagsResponse:
      description: Response body returned by all three tag management endpoints.
      type: object
      required: [kid, tags]
      properties:
        kid:
          type: string
          description: KMS unique identifier of the key.
...
paths:
...
  /.well-known/jwks.json:
    get:
      tags: [JWKS]
      summary: JSON Web Key Set
      description: |
        Returns the public key set (RFC 7517) for all KMS keys that carry the
        `"jwks"` tag and are in Active or Deactivated state.

        This endpoint is **intentionally unauthenticated** — no credentials are
        required. Enable it only when you want third parties (e.g. JWT verifiers)
        to discover the server's public keys.

        When the response would exceed `jwks_endpoint_max_keys` entries, the set
        is truncated and the header `X-JWKS-Truncated: true` is added.
      operationId: getJwks
      security: []
      responses:
        '200':
          description: JWK Set
          content:
            application/json:
              schema:
                type: object
                required: [keys]
                properties:
                  keys:
                    type: array
                    description: Array of public JWK objects
                    items:
                      type: object
                      required: [kty, kid]
                      properties:
                        kty:
                          type: string
                          description: Key type (EC, RSA, OKP)
                          example: EC
                        kid:
                          type: string
                          description: KMS unique identifier
                        alg:
                          type: string
                          description: JOSE algorithm
                        use:
                          type: string
                          description: Public key use (sig or enc)
                        key_ops:
                          type: array
                          items:
                            type: string
                          description: Permitted key operations
                        crv:
                          type: string
                          description: Elliptic curve / OKP curve name
                        x:
                          type: string
                          description: EC/OKP x-coordinate (base64url)
                        y:
                          type: string
                          description: EC y-coordinate (base64url)
                        n:
                          type: string
                          description: RSA modulus (base64url)
                        e:
                          type: string
                          description: RSA public exponent (base64url)
              example:
                keys:
                  - kty: EC
                    kid: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    alg: ES256
                    use: sig
                    crv: P-256
                    x: f83OJ3D2xF1Bg8vub9tLe1gHMzV76e8Tus9uPHvRVEU
                    y: x_FEzRu9m36HLN_tue659LNpXW6pCyStikYjKIWI5a0
        '500':
          description: Internal server error

  # ── Vault Transit ──────────────────────────────────────────────────────────
  # Enabled only when vault_api_enabled = true in server config.
  # The {transit_mount} defaults to "transit" but is configurable.
  # All endpoints require X-Vault-Token validated against auth-verifier.
...